feat(futureboxes): futurebox 조회 페이지 완성

+++ src/main/java/com/stdev/futurebox/controller/FuturePerfumeController.java
@@ -9,6 +9,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 @Slf4j
 @Controller
@@ -37,12 +38,34 @@ public class FuturePerfumeController {
     public String perfume(@PathVariable Long id, Model model) {
         FuturePerfume perfume = futurePerfumeService.findById(id);
         model.addAttribute("perfume", perfume);
-        return "future-perfume/perfume";
+        return "future-perfume/futurePerfume";
     }
 
-    @PostMapping("/delete")
-    public String delete(@RequestParam Long id, @RequestParam Long boxId) {
+    @PostMapping("/delete/{id}")
+    public String delete(@PathVariable Long id, RedirectAttributes redirectAttributes) {
+        FuturePerfume perfume = futurePerfumeService.findById(id);
+        if (perfume == null) {
+            throw new IllegalArgumentException("해당 ID의 FuturePerfume이 존재하지 않습니다: " + id);
+        }
+        
+        Long boxId = perfume.getBoxId();
         futurePerfumeService.delete(id);
+        redirectAttributes.addFlashAttribute("message", "Future Perfume deleted successfully.");
         return "redirect:/future-boxes/" + boxId;
+    }
+
+    @GetMapping("/edit/{id}")
+    public String editForm(@PathVariable Long id, Model model) {
+        FuturePerfume perfume = futurePerfumeService.findById(id);
+        model.addAttribute("perfume", perfume);
+        return "future-perfume/editPerfumeForm";
+    }
+
+    @PostMapping("/edit/{id}")
+    public String edit(@PathVariable Long id, @ModelAttribute FuturePerfumeCreateForm form) {
+        FuturePerfume perfume = form.toEntity();
+        perfume.setId(id);
+        futurePerfumeService.update(perfume);
+        return "redirect:/future-perfume/" + perfume.getId();
     }
 }
